* EDM-5: crud functionality implemented, patch added * EDM-5: UI is divided into components, improving business logic
