+++ RELEASE_NOTES.md
@@ -1,0 +1,18 @@
+# Release Notes
+
+## 0.4.0-SNAPSHOT
+
+* EDM-5: crud functionality implemented, RELEASE_NOTES.md added
+* EDM-5: UI is divided into components, improving business logic
+
+## 0.3.0-SNAPSHOT
+
+* EDM-4: improved function for updating and reading data from Excel
+
+## 0.2.0-SNAPSHOT
+
+* EDM-2: Excel table reading finally fixed
+
+## 0.1.0-SNAPSHOT
+
+* EDM-1: transferring a project from local access to GitHub.
